Update: 検索処理を別ファイルに分割（xcode 13.0）
