+++ ViewCotrollerSearch.swift
@@ -1,0 +1,8 @@
+//
+//  ViewCotrollerSearch.swift
+//  GpsMap_2
+//
+//  Created by hiroto on 2021/09/28.
+//
+
+import Foundation
